requirements.txt
Flask==2.2.2
openai==1.3.0
pandas==2.1.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.20
pandasql==0.7.3
python-dotenv==1.0.0
